Populating CRAM cache is now much less likely to have a race condition.
On filesystems that don't support exclusive fopen, it was possible
totrigger a race condition causing the downloaded data to become
corrupt.  We now randomise the filenames to make this much less
likely.

The data is also md5sumed before writing.  (Logically we'd check after
writing, but if we've hit an NFS issue with multiple clients writing
to the same file then we're unlikely to detect that anyway as each
system will just read back what it wrote from its own local cache.)